moved unit tests to unit package
