+++ src/test/java/br/ufpb/dcx/apps4society/educapi/unit/service/JWTServiceTest.java
@@ -1,12 +1,12 @@
-package br.ufpb.dcx.apps4society.educapi.service;
+package br.ufpb.dcx.apps4society.educapi.unit.service;
 
 import br.ufpb.dcx.apps4society.educapi.domain.User;
-import br.ufpb.dcx.apps4society.educapi.domain.builder.UserBuilder;
 import br.ufpb.dcx.apps4society.educapi.dto.user.UserLoginDTO;
 import br.ufpb.dcx.apps4society.educapi.repositories.UserRepository;
 import br.ufpb.dcx.apps4society.educapi.response.LoginResponse;
 import br.ufpb.dcx.apps4society.educapi.services.JWTService;
 import br.ufpb.dcx.apps4society.educapi.services.exceptions.InvalidUserException;
+import br.ufpb.dcx.apps4society.educapi.unit.domain.builder.UserBuilder;
 import br.ufpb.dcx.apps4society.educapi.util.Messages;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -18,25 +18,24 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.util.ReflectionTestUtils;
 
 import java.util.Optional;
+
 import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith(MockitoExtension.class)
 public class JWTServiceTest {
 
-    private UserLoginDTO userLoginDTO = UserBuilder.anUser().buildUserLoginDTO();
-    private Optional<User> userOptional = UserBuilder.anUser().buildOptionalUser();
-    private String invalidToken = "Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJtYWlhd2VlZUB0ZXN0LmNvbSIsImV4cCI6MTYxNTM" +
+    private final UserLoginDTO userLoginDTO = UserBuilder.anUser().buildUserLoginDTO();
+    private final Optional<User> userOptional = UserBuilder.anUser().buildOptionalUser();
+    private final String invalidToken = "Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJtYWlhd2VlZUB0ZXN0LmNvbSIsImV4cCI6MTYxNTM" +
             "5OTkyN30.1qNJIgwjlnm6YcZuIDFLZrQLs58qOwLFkCtXOcaUD-fQZyTa4usOMVgGa19Em_e8WdoXfnaJSv9O-c8IRp-C9Q";
+    @Mock
+    UserRepository userRepository;
+    @InjectMocks
+    JWTService service;
 
     private String tokenFormat(String token) {
         return "Bearer " + token;
     }
-
-    @Mock
-    UserRepository userRepository;
-
-    @InjectMocks
-    JWTService service;
 
     @BeforeEach
     public void setUp() {
